+++ packages/fluent_sdk/lib/src/api/registry.dart
@@ -30,11 +30,25 @@ abstract class Registry {
   });
 
   /// Checks if a type [T] is currently registered in the container.
-  bool isRegistered<T extends Object>();
+  bool isRegistered<T extends Object>({String? instanceName});
 
   /// Retrieves an instance of a registered object [T].
   T get<T extends Object>({String? instanceName});
 
   /// Retrieves an instance of a registered object [T] (shorthand for [get]).
   T call<T extends Object>({String? instanceName});
+
+  /// Resets a Lazy Singleton, allowing it to be recreated on the next request.
+  void resetLazySingleton<T extends Object>({
+    T? instance,
+    String? instanceName,
+    void Function(T)? disposingFunction,
+  });
+
+  /// Unregisters an instance of a registered object [T].
+  void unregister<T extends Object>({
+    T? instance,
+    String? instanceName,
+    void Function(T)? disposingFunction,
+  });
 }