+++ packages/fluent_localization/fluent_localization/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.9.0
+
+* CHORE: Updated `fluent_sdk` dependency to `^0.8.0`.
+
 ## 1.8.1
 
 * PERF: Optimized recursive string flattening and map processing.